sessions: release datatypes without checking cookie
Turns out that the predefined datatypes are initialized statically,
so the first time through calling ompi_datatype_finalize the magic
cookie gets cleared if code is compiled in debug mode.  This results
in a second call to ompi_datatype_finalize to blow up with an assert
since the cookie had been cleared in the first pass.

Add a new OBJ_DESTRUCT variant which does not check/clear cookie even
when the code is compiled in --enable-debug mode.

Signed-off-by: Howard Pritchard <howardp@lanl.gov>